requirements.txt
fastapi
uvicorn[standard]
jinja2
aiosqlite
numpy
pandas
xlsxwriter
python-dotenv
sqlalchemy
plotly
kaleido
scipy